The source files are now entirely ASCII.  This should allow usability by those who cannot properly display or efficiently type the Unicode symbols that were previously used.  A separate bindings-library will soon be provided to allow users to use these Unicode symbols if they wish.
The "label" of a State and the "path" of a Transition are now referred to as "nodeLabel" and "edgeLabel", respectively.  Further, it turns out that our "SyntacticMonoid" was not actually a syntactic monoid, so it has been renamed "PowerSetGraph".

My implementation of the SL-check has been known to be incorrect essentially since its introduction, but it was only now removed.  A correct and extended implementation will be provided shortly.

Aside from these renamings and removals, no functionality changes have occurred.