In r20, states were given amnesia regarding their contained values.  There are pros and cons to this approach, but we have since decided that it is the wrong one.  Some things that have changed:
* An FSA is no longer a Container
* (==) on FSAs no longer works when their nodeLabel types differ
* A function `isomorphic` has been added to allow this
* Operations on States no longer have any silliness involving Show
* Type signatures changed almost everywhere

Almost everything worked as expected after merely changing the type signature to refect the new information, but I had done something strange in combineTransitions that prevented that function from compiling with a reasonable type.  It has been rewritten in a less-strange way.  Determinization was previously a no-op on already-deterministic FSAs, but with the new information it has to do something to conform to the appropriate type.  It simply wraps every nodeLabel as a singleton Set in such cases.