The type arguments to FSA now list the nodeLabel type first, so a lot of source lines (type signatures) changed.  Again.  Also, there is now a complementDeterminized function which throws an error if its input is nondeterministic.  The utility of functions that result in runtime errors when valid (but slower) solutions exist will be discussed soon.  I don't like it.
An FSA with nodes labelled by an enumerated type with a total order is now able to act as a container (though arbitrary FSAs are not).  I am not completely certain this is useful, but for now I believe it is.  The restriction on nodeLabel type is to satisfy certain requirements of the type system, not a reasonable semantic restriction.

LogicClasses.lhs now defines a class "Linearizable" for structures that can be traversed an element at a time.  Using reasonable semantic notions for the provided methods, anything that is Collapsible must also be Linearizable, but I would imagine that an inifinite generator class, perhaps of pseudo-random numbers, might be Linearizable but not Collapsible.

The functions in readJeff.lhs are no longer monadic in their return types, which appears to have simplified much of the code.