Mostly intra-module reorganization.
Changed makePairs to a functionally identical one using partial application in place of one lambda expression.  Honestly, it is just the Cartesian product of two list-like things, so written with a different type-signature could be applied to any collapsible containers.  I may move it over to LogicClasses (with an appropriate signature) due to that fact.

A few descriptive comments have been added, clarifying correctness and preconditions of algorithms.