Syntactic monoids and true isomorphism
The isomorphism test only works for DFAs, as each state may have several out-edges on a single symbol and it is in general difficult to map which destination state corresponds to which.

Syntactic monoids are important for the piecewise tests, but those are not yet implemented.  The next updates will be progress toward those.